+++ lib/definition/dsl.rb
@@ -89,5 +89,11 @@ module Definition
                       end
                     ))
     end
+
+    # Example:
+    # Nilable(Definition.Type(Integer))
+    def Nilable(definition) # rubocop:disable Style/MethodName
+      Types::Or.new(:nullable, Nil(), definition)
+    end
   end
 end