+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.5.0] - 2019-03-28
 ### Added
 - CoercibleValueObject Definition for better nesting of ValueObjects
+- Nilable Definition as shortcut for nil OR some other definition
 - Option for Keys Definition to ignore unexpected keys
 ### Fixed
 - Error hash was missing some errors in a few cases